fix: clear cookie when the avatar has been deleted
This patch clears cookie when the avatar has been deleted, which happens
when users upload a new avatar, the origial avatar on gcs would be deleted
therefore the corresponding value of 'keystone.avatar' in cookie should
be cleared at the same time.

Also, it fixes the 'delete' schema method to configure it with correct
gcs config and data path.